refactor: move hamburger menu to sepearated file
